+++ core/src/com/trance/common/socket/handler/ResponseProcessor.java
@@ -1,14 +1,14 @@
 package com.trance.common.socket.handler;
 
-import com.trance.common.socket.model.Response;
-
 import org.apache.mina.core.session.IoSession;
+
+import com.trance.common.socket.model.Response;
 
 
 /**
  * 响应消息处理器
  * 
- * @author zhangyl
+ * @author trance
  */
 public interface ResponseProcessor {
 	
@@ -16,17 +16,18 @@ public interface ResponseProcessor {
 	 * 获取模块号
 	 * @return
 	 */
-	int getModule();
+	byte getModule();
 	
 	/**
 	 * 获取命令号
 	 * @return
 	 */
-	int getCmd();
+	byte getCmd();
 	
 	/**
 	 * 响应消息的对象类型, null-业务自己做转换
-	 * @return Object
+	 * @param <T>
+	 * @return Class<?> 基本类型暂时只技能 Integer, Long , String 
 	 */
 	Class<?> getType();
 	
@@ -36,5 +37,5 @@ public interface ResponseProcessor {
 	 * @param response {@link Response}
 	 */
 	void callback(IoSession session, Response response);
-	
+
 }
