+++ core/src/com/trance/empire/modules/dailyreward/handler/DailyRewardCmd.java
@@ -2,5 +2,5 @@ package com.trance.empire.modules.dailyreward.handler;
 
 public interface DailyRewardCmd {
 	
-	int GET_DAILY_REWARD = 1;
+	byte GET_DAILY_REWARD = 1;
 }
