+++ core/src/com/trance/empire/modules/replay/handler/ReplayCmd.java
@@ -2,5 +2,5 @@ package com.trance.empire.modules.replay.handler;
 
 public interface ReplayCmd {
 	
-	int GET_PLAYER_REPLAY = 1;
+	byte GET_PLAYER_REPLAY = 1;
 }
